requirements.txt
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
tensorflow==2.15.0
streamlit
openpyxl
pillow